# latency: disable heavy precompilation