--- a/src/ImageCore.jl
+++ b/src/ImageCore.jl
@@ -170,9 +170,9 @@ function Base.transpose(a::AbstractVector{C}) where C<:Colorant
     out
 end
 
-if VERSION >= v"1.4.2" # work around https://github.com/JuliaLang/julia/issues/34121
-    include("precompile.jl")
-    _precompile_()
-end
+# if VERSION >= v"1.4.2" # work around https://github.com/JuliaLang/julia/issues/34121
+#     include("precompile.jl")
+#     _precompile_()
+# end
 
 end ## module